cleanup

+++ daemon/extension/CommandScript.h
@@ -54,8 +54,8 @@ public:
 private:
 	MessageList m_messages;
 	Mutex m_logMutex;
-	int64 m_idMessageGen;
-	int64 m_idScriptGen;
+	int m_idMessageGen;
+	int m_idScriptGen;
 };
 
 extern CommandScriptLog* g_CommandScriptLog;
